projects/stanton-times/requirements.txt
# Core dependencies
requests>=2.25.1
feedparser>=6.0.0
beautifulsoup4>=4.9.3
discord.py>=2.3.0
python-dotenv>=0.15.0
numpy>=1.23.0
scikit-learn>=1.3.0

# Logging
structlog>=21.1.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0
mock>=4.0.3